requirements.txt
Django==5.1.3
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pillow==11.0.0
django-filter==24.3
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
stripe==11.3.0
celery==5.4.0
redis==5.2.1
